Update vendor files
The Backbone library is no longer needed. Include the latest versions of
Ampersand Model and Ampersand Collection for equivalent testing
purposes. These modules are intended to be consumed with Browserify, so
their `node_modules` directories must be forecibly added to the project.
Include a `.gitattributes` file labeling all files within the `vendor`
directory as binary to better reflect their purpose.

These changes are encapsulated in a single commit in order to simplify
history review.